+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -1,0 +1,23 @@
+package school.faang.user_service.controller.event;
+
+import lombok.RequiredArgsConstructor;
+
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import school.faang.user_service.service.event.EventParticipationService;
+
+@RequestMapping("/participation")
+@RestController
+@RequiredArgsConstructor
+public class EventParticipationController {
+    private final EventParticipationService eventParticipationService;
+
+    @PostMapping
+    public void registerParticipant() {
+        System.out.println("registerParticipant");
+    }
+
+
+}
