--- src/main/java/school/faang/user_service/dto/EventDto.java
@@ -1,13 +1,0 @@
-package school.faang.user_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class EventDto {
-    private long id;
-    private String name;
-}
