+++ src/main/java/school/faang/user_service/dto/event/participant/EventDto.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.dto.event.participant;
+
+public class EventDto {
+}
