+++ src/main/java/school/faang/user_service/dto/event/participant/RegisterParticipantDto.java
@@ -1,10 +1,12 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.event.participant;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@NotNull
 @NoArgsConstructor
 @AllArgsConstructor
 public class RegisterParticipantDto {
