+++ src/main/java/school/faang/user_service/service/event/EventParticipationService.java
@@ -1,11 +1,10 @@
 package school.faang.user_service.service.event;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.RegisterParticipantDto;
-import school.faang.user_service.exception.DataValidationException;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.GetMapping;
+import school.faang.user_service.dto.event.participant.RegisterParticipantDto;
 import school.faang.user_service.repository.event.EventParticipationRepository;
 
 import java.util.HashMap;
@@ -14,14 +13,22 @@ import java.util.Map;
 @Service
 @RequiredArgsConstructor
 public class EventParticipationService {
-    private final Map<Long, Long> eventParticipationMap = new HashMap<>();
     private final EventParticipationRepository eventParticipationRepository;
+    private final Map<Long, Long> particaptionDtoMap = new HashMap<>();
 
+    @Transactional
+    public void registerParticipant(RegisterParticipantDto registerParticipantDto) {
+        long eventId = registerParticipantDto.getEventId();
+        long userId = registerParticipantDto.getUserId();
 
-    public void registerParticipant(long eventId, long userId) {
-        if (eventParticipationMap.get(eventId) == null || eventParticipationMap.get(userId) == null) {
-            throw new DataValidationException("ID не может быть пустым!");
-        }
-        eventParticipationMap.put(eventId, userId);
+        particaptionDtoMap.put(eventId, userId);
+        eventParticipationRepository.register(particaptionDtoMap.get(eventId), userId);
     }
+
+    @Transactional
+    EventParticipationService eventParticipationService() {
+        return new EventParticipationService(eventParticipationRepository);
+    }
+
+
 }
